weird old changes
check it does not screw things up

+++ src/matlab/Processing/plot_CAP_w.m
@@ -35,7 +35,7 @@ EEG=pop_loadbv;
  tau=100;
  size_bin=floor(tau*Fs/1000);
  
- Data= EEG.data';
+ Data= double(EEG.data');
  
  %Data=detrend(Data);
